# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  github_com_xilo-platform_xilo_internal_auth_model.User:
    properties:
      avatar_url:
        type: string
      bio:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      display_name:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      id:
        type: string
      phone:
        type: string
      role:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
...
  github_com_xilo-platform_xilo_internal_comment_model.Comment:
    properties:
      author:
        $ref: '#/definitions/github_com_xilo-platform_xilo_internal_auth_model.User'
      author_id:
        type: string
      content:
        type: string
      content_html:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      depth:
        type: integer
      id:
        type: string
      is_pinned:
        type: boolean
      is_spam:
        type: boolean
      media_url:
        type: string
      parent_id:
        type: string
      post_id:
        type: string
      reactions:
        additionalProperties:
          type: integer
        type: object
      replies:
        items:
          $ref: '#/definitions/github_com_xilo-platform_xilo_internal_comment_model.Comment'
        type: array
      root_id:
        type: string
      updated_at:
        type: string
    type: object
  github_com_xilo-platform_xilo_internal_comment_model.CreateCommentRequest:
    properties:
      content:
        type: string
      media_url:
        type: string
      parent_id:
        type: string
      root_id:
        type: string
    type: object
  github_com_xilo-platform_xilo_internal_media_model.Media:
    properties:
      created_at:
        type: string
      filename:
        type: string
      height:
        type: integer
      id:
        type: string
      mime_type:
        type: string
      original_name:
        type: string
      size_bytes:
        type: integer
      user_id:
        type: string
      variants:
        additionalProperties:
          type: string
        type: object
      width:
        type: integer
    type: object
  github_com_xilo-platform_xilo_internal_media_model.UploadResponse:
    properties:
      height:
        type: integer
      id:
        type: string
      size:
        type: integer
      url:
        type: string
      variants:
        additionalProperties:
          type: string
        type: object
      width:
        type: integer
    type: object
...
host: localhost:8000
info:
  contact:
    email: support@xilo.example.com
    name: Xilo API Support
    url: https://xilo.example.com/support
  description: A self-hosted, Telegram-inspired modern blog platform API.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://xilo.example.com/terms
  title: Xilo Blog Platform API
  version: "1.0"
paths:
  /{type}/{id}/reactions:
    post:
      consumes:
      - application/json
      parameters:
      - description: Target type (post/comment)
        in: path
        name: type
        required: true
        type: string
      - description: Target ID
        in: path
        name: id
        required: true
        type: string
      - description: Reaction data
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Toggle reaction on a target
      tags:
      - reactions
...
  /billing/settings:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get payment gateway config
      tags:
      - billing
    patch:
      consumes:
      - application/json
      parameters:
      - description: Config data
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update payment gateway config
      tags:
      - billing
...
  /comments/{id}/pin:
    post:
      consumes:
      - application/json
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: string
      - description: Pin status
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Pin/unpin a comment
      tags:
      - comments
...
  /media:
    get:
      parameters:
      - description: Pagination cursor
        in: query
        name: cursor
        type: string
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List user media
      tags:
      - media
  /media/{id}:
    delete:
      parameters:
      - description: Media ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete media
      tags:
      - media
    get:
      parameters:
      - description: Media ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_xilo-platform_xilo_internal_media_model.Media'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get media by ID
      tags:
      - media
  /media/upload:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: File to upload
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_xilo-platform_xilo_internal_media_model.UploadResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Upload media file
      tags:
      - media
...
  /posts/{postId}/comments:
    get:
      parameters:
      - description: Post ID
        in: path
        name: postId
        required: true
        type: string
      - description: Pagination cursor
        in: query
        name: cursor
        type: string
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      - default: newest
        description: Sort order (newest/oldest)
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: List comments for a post
      tags:
      - comments
    post:
      consumes:
      - application/json
      parameters:
      - description: Post ID
        in: path
        name: postId
        required: true
        type: string
      - description: Comment data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_xilo-platform_xilo_internal_comment_model.CreateCommentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_xilo-platform_xilo_internal_comment_model.Comment'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a comment
      tags:
      - comments
  /posts/{postId}/comments/{id}:
    delete:
      parameters:
      - description: Post ID
        in: path
        name: postId
        required: true
        type: string
      - description: Comment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a comment
      tags:
      - comments
    patch:
      consumes:
      - application/json
      parameters:
      - description: Post ID
        in: path
        name: postId
        required: true
        type: string
      - description: Comment ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated comment
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_xilo-platform_xilo_internal_comment_model.Comment'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a comment
      tags:
      - comments
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"